LayoutTests/css3/calc/margin-expected.txt

PASS computedMarginLeft("simple-all") is "25px"
PASS computedMarginTop("simple-all") is "25px"
PASS computedMarginRight("simple-all") is "25px"
PASS computedMarginBottom("simple-all") is "25px"
PASS computedMarginLeft("simple-left") is "25px"
PASS computedMarginTop("simple-left") is "0px"
PASS computedMarginRight("simple-left") is "0px"
PASS computedMarginBottom("simple-left") is "0px"
PASS computedMarginLeft("simple-right") is "0px"
PASS computedMarginTop("simple-right") is "0px"
PASS computedMarginRight("simple-right") is "25px"
PASS computedMarginBottom("simple-right") is "0px"
PASS computedMarginLeft("simple-top") is "0px"
PASS computedMarginTop("simple-top") is "25px"
PASS computedMarginRight("simple-top") is "0px"
PASS computedMarginBottom("simple-top") is "0px"
PASS computedMarginLeft("simple-bottom") is "0px"
PASS computedMarginTop("simple-bottom") is "0px"
PASS computedMarginRight("simple-bottom") is "0px"
PASS computedMarginBottom("simple-bottom") is "25px"
FAIL computedMarginLeft("percent-all") should be 25px. Was 0px.
FAIL computedMarginTop("percent-all") should be 25px. Was 0px.
FAIL computedMarginRight("percent-all") should be 25px. Was 0px.
FAIL computedMarginBottom("percent-all") should be 25px. Was 0px.
FAIL computedMarginLeft("percent-left") should be 25px. Was 0px.
PASS computedMarginTop("percent-left") is "0px"
PASS computedMarginRight("percent-left") is "0px"
PASS computedMarginBottom("percent-left") is "0px"
PASS computedMarginLeft("percent-right") is "0px"
PASS computedMarginTop("percent-right") is "0px"
FAIL computedMarginRight("percent-right") should be 25px. Was 0px.
PASS computedMarginBottom("percent-right") is "0px"
PASS computedMarginLeft("percent-top") is "0px"
FAIL computedMarginTop("percent-top") should be 25px. Was 0px.
PASS computedMarginRight("percent-top") is "0px"
PASS computedMarginBottom("percent-top") is "0px"
PASS computedMarginLeft("percent-bottom") is "0px"
PASS computedMarginTop("percent-bottom") is "0px"
PASS computedMarginRight("percent-bottom") is "0px"
FAIL computedMarginBottom("percent-bottom") should be 25px. Was 0px.
PASS successfullyParsed is true

TEST COMPLETE
